chore(nlp): Making ONNX-based scripts ready to be used.
