+++ archai/nlp/common/constants.py
@@ -1,6 +1,8 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-"""Provides post-exported ONNX optimization
-classes and methods.
+"""Constants definitions used within Archai.
 """
+
+BATCH_SIZE = 1
+SEQ_LEN = 32
