+++ archai/nlp/compression/onnx/onnx_utils/export.py
@@ -21,6 +21,7 @@ def weight_sharing(onnx_model_path: str, model_type: str) -> None:
 
     Args:
         onnx_model_path: Path to the ONNX model that will have weights shared.
+        model_type: Type of model to share the weights.
 
     """
 
@@ -42,17 +43,17 @@ def weight_sharing(onnx_model_path: str, model_type: str) -> None:
     if model_type == 'hf_gpt2':
         n_emb_weight = 1
         n_cutoffs = 0
+
     elif model_type == 'mem_transformer':
         n_emb_weight = len(list(filter(lambda x: 'word_emb.emb_layers' in x, weights.keys())))
         n_cutoffs = n_emb_weight - 1
+
     else:
         raise ValueError(f'Model {model_type} not supported for weight sharing.')
-
 
     for i in range(n_emb_weight):
         # Grabs the embedding weights pointer and removes from the graph
         emb_weight_name = f'word_emb.emb_layers.{i}.weight'
-
         if model_type == 'hf_gpt2':
             emb_weight_name = 'transformer.wte.weight'
 
@@ -85,13 +86,13 @@ def weight_sharing(onnx_model_path: str, model_type: str) -> None:
     save(model, onnx_model_path)
 
 
-def export_onnx_from_pt(model: torch.nn.Module,
-                        model_config: dict,
-                        model_type: str,
-                        onnx_model_path: str,
-                        share_weights: Optional[bool] = True,
-                        do_constant_folding: Optional[bool] = True,
-                        opset_version: Optional[int] = 11) -> None:
+def export_onnx_from_torch(model: torch.nn.Module,
+                           model_config: dict,
+                           model_type: str,
+                           onnx_model_path: str,
+                           share_weights: Optional[bool] = True,
+                           do_constant_folding: Optional[bool] = True,
+                           opset_version: Optional[int] = 11) -> None:
     """Exports a PyTorch-based model to ONNX.
 
     Args:
@@ -107,8 +108,6 @@ def export_onnx_from_pt(model: torch.nn.Module,
 
     # Gathers the proper ONNX configuration instance
     onnx_config = load(model_type, cls_type='config', model_config=model_config)
-
-    print(onnx_config)
 
     # Creates the dynamic axes based on inputs and outputs
     dynamic_axes = {name: axes for name, axes in chain(onnx_config.inputs.items(), onnx_config.outputs.items())}
