+++ archai/nlp/compression/onnx/onnx_utils/optimization.py
@@ -1,6 +1,9 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
+"""ONNX-related optimization helpers and utilities.
+"""
+
 from pathlib import Path
 from typing import Optional
 
@@ -8,7 +11,7 @@ from onnx import load_model
 from onnxruntime.transformers.optimizer import optimize_by_onnxruntime
 
 from archai.nlp.common.file_naming_utils import create_file_name_identifier
-from archai.nlp.compression.onnx.onnx_utils.opt.fusion_options import FusionOptions
+from archai.nlp.compression.onnx.onnx_utils.fusion_options import FusionOptions
 from archai.nlp.common.lazy_loader import load
