+++ archai/nlp/compression/quantization/ptq.py
@@ -16,7 +16,7 @@ from onnxruntime.quantization.quant_utils import attribute_to_kwarg, ms_domain
 from onnxruntime.quantization.quantize import quantize_dynamic
 from onnxruntime.quantization.registry import IntegerOpsRegistry
 
-from archai.nlp.common.lazy_loader import load
+from archai.nlp.common.lazy_loader import load_from_checkpoint
 from archai.nlp.compression.onnx.onnx_utils.load import create_file_name_identifier
 
 
@@ -151,8 +151,9 @@ def dynamic_quantization_torch(torch_model_path: str,
     torch.set_num_threads(1)
 
     # Loads the pre-trained model
-    model = load(model_type, cls_type='model')
-    model.update_with_checkpoint(torch_model_path, on_cpu=True)
+    model = load_from_checkpoint(model_type,
+                                 torch_model_path,
+                                 on_cpu=True)
 
     # Performs an initial dynamic quantization
     model_qnt = torch.quantization.quantize_dynamic(model, {torch.nn.Linear})
